level: extract loader module
