+++ src/libtrx/include/libtrx/game/level.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "level/common.h"
+#include "level/loader.h"
 #include "level/reader.h"
 #include "level/settings.h"
 #include "level/types.h"
